Rename modules and expect sluggified inputs
