+++ dtocean_reliability/parse.py
@@ -25,16 +25,12 @@ DTOcean Reliability Assessment Module (RAM)
 """
 
 # Built in modules
+import re
 import logging
-from copy import copy, deepcopy
+from copy import deepcopy
 from collections import Counter, OrderedDict
 
-from .data import (Component,
-                   Parallel,
-                   ReliabilityWrapper,
-                   Serial,
-                   find_all_labels,
-                   find_strings)
+from .graph import Component, Parallel, Serial
 
 # Start logging
 module_logger = logging.getLogger(__name__)
@@ -50,257 +46,6 @@ class SubNetwork(object):
         return
 
 
-class Network(object):
-    
-    def __init__(self, database,
-                       electrical_network = None,
-                       moorings_network = None,
-                       user_network = None,
-                       electrical_data = None):
-        
-        if (electrical_network is None and
-            moorings_network is None and
-            user_network is None):
-            
-            err_msg = "At least one network input must be provided"
-            raise ValueError (err_msg)
-        
-        _check_nodes(electrical_network, moorings_network)
-        
-        (electrical_network,
-         moorings_network,
-         user_network) = _complete_networks(electrical_network,
-                                            moorings_network,
-                                            user_network)
-        
-        (array_hierarcy,
-         device_hierachy) = _combine_networks(electrical_network,
-                                              moorings_network,
-                                              user_network,
-                                              electrical_data)
-        
-        self._db = database
-        self._pool = _build_pool(array_hierarcy, device_hierachy)
-        self._subhub_indices = _get_indices(self._pool, "subhub")
-        self._device_indices = _get_indices(self._pool, "device")
-        self._curtailments = _get_curtailments(self._pool)
-        self._system_root = ["device", "subhub", "array"]
-    
-    def set_failure_rates(self, severitylevel='critical',
-                                calcscenario='mean',
-                                use_kfactors=False,
-                                inplace=False):
-        
-        if inplace:
-            network = self
-        else:
-            network = copy(self)
-            network._pool = deepcopy(self._pool)
-        
-        _set_component_failure_rates(network._pool,
-                                     network._db,
-                                     severitylevel,
-                                     calcscenario,
-                                     use_kfactors=use_kfactors)
-        
-        if inplace:
-            result = None
-        else:
-            result = network
-        
-        return result
-    
-    def get_systems_metrics(self, time_hours=None):
-        
-        indices = []
-        systems = []
-        failure_rates = []
-        mttfs = []
-        rpns = []
-        reliabilities = []
-        
-        # Array
-        array = self._pool["array"]
-        indices.append("array")
-        systems.append("array")
-        failure_rates.append(array.get_failure_rate(self._pool))
-        mttfs.append(array.get_mttf(self._pool))
-        rpns.append(array.get_rpn(self._pool))
-        
-        if time_hours is not None:
-            reliabilities.append(array.get_reliability(self._pool, time_hours))
-        
-        # Subhub
-        if self._subhub_indices is not None:
-            
-            subhub_names = sorted(self._subhub_indices.keys())
-            
-            for name in subhub_names:
-                
-                idx = self._subhub_indices[name]
-                subhub = self._pool[idx]
-                
-                indices.append(idx)
-                systems.append(name)
-                failure_rates.append(subhub.get_failure_rate(self._pool))
-                mttfs.append(subhub.get_mttf(self._pool))
-                rpns.append(subhub.get_rpn(self._pool))
-                
-                if time_hours is not None:
-                    reliabilities.append(subhub.get_reliability(self._pool,
-                                                                time_hours))
-        
-        # Devices
-        if self._device_indices is not None:
-            
-            device_names = sorted(self._device_indices.keys())
-            
-            for name in device_names:
-                
-                idx = self._device_indices[name]
-                device = self._pool[idx]
-                
-                indices.append(idx)
-                systems.append(name)
-                failure_rates.append(device.get_failure_rate(self._pool))
-                mttfs.append(device.get_mttf(self._pool))
-                rpns.append(device.get_rpn(self._pool))
-                
-                if time_hours is not None:
-                    reliabilities.append(device.get_reliability(self._pool,
-                                                                time_hours))
-        
-        if set(failure_rates) == set([None]): return None
-        
-        result = OrderedDict()
-        result["Link"] = indices
-        result["System"] = systems
-        result["lambda"] = failure_rates
-        result["MTTF"] = mttfs
-        result["RPN"] = rpns
-    
-        if time_hours is not None:
-            key = "R ({} hours)".format(time_hours)
-            result[key] = reliabilities
-        
-        return result
-    
-    def get_subsystem_metrics(self, subsystem_name, time_hours=None):
-        
-        def get_lowest_system(labels):
-            
-            for system in self._system_root:
-                for label in labels:
-                    if system in label:
-                        return label
-            
-            raise ValueError("No system found")
-        
-        self._check_not_system(subsystem_name)
-    
-        all_labels, indices = find_all_labels(subsystem_name, self._pool)
-        
-        if all_labels is None: return None
-        
-        systems = []
-        failure_rates = []
-        mttfs = []
-        rpns = []
-        reliabilities = []
-        
-        for labels, index in zip(all_labels, indices):
-            
-            link = self._pool[index]
-            
-            systems.append(get_lowest_system(labels))
-            failure_rates.append(link.get_failure_rate(self._pool))
-            mttfs.append(link.get_mttf(self._pool))
-            rpns.append(link.get_rpn(self._pool))
-            
-            if time_hours is not None:
-                reliabilities.append(link.get_reliability(self._pool,
-                                                          time_hours))
-        
-        # Build curtailments
-        curtailments = []
-        
-        for system in systems:
-            
-            if system == "array" or "subhub" in system:
-                curtailments.append(self._curtailments[system])
-                continue
-            
-            if subsystem_name in ("Array elec sub-system",
-                                  "Elec sub-system"):
-                
-                curtailments.append(self._curtailments[system])
-                continue
-            
-            curtailments.append([system])
-        
-        if set(failure_rates) == set([None]): return None
-
-        result = OrderedDict()
-        result["Link"] = indices
-        result["System"] = systems
-        result["lambda"] = failure_rates
-        result["MTTF"] = mttfs
-        result["RPN"] = rpns
-        
-        if time_hours is not None:
-            key = "R ({} hours)".format(time_hours)
-            result[key] = reliabilities
-        
-        result["Curtails"] = curtailments
-        
-        return result
-    
-    def display(self):
-        return self._pool['array'].display(self._pool)
-    
-    def _check_not_system(self, name):
-                
-        if any([x in name for x in self._system_root]):
-            
-            reserved_str = ", ".format(self._system_root)
-            err_str = ("Subsystem name may not contain reserved keywords: "
-                       "{}").format(reserved_str)
-            raise ValueError(err_str)
-    
-    def _find_system_index(self, system_name):
-        
-        try:
-            
-            _, index = find_all_labels(system_name,
-                                       self._pool,
-                                       return_one=True)
-        
-        except RuntimeError:
-            
-            try:
-                _, index = find_all_labels(system_name,
-                                           self._pool,
-                                           return_shortest=True)
-            except RuntimeError:
-                err_str = "No unique subsystem failure rate could be found"
-                raise RuntimeError(err_str)
-        
-        return index
-    
-    def __getitem__(self, key):
-        return ReliabilityWrapper(self._pool, key)
-    
-    def __len__(self):
-        
-        result = 0
-        
-        for s in self._pool:
-            if self._pool[s].label is not None:
-                result += 1
-        
-        return result
-
-
 class KSystem(object):
     
     def __init__(self, ids, kfactors):
@@ -314,7 +59,18 @@ class KSystem(object):
         return self.__str__()
 
 
-def _check_nodes(*networks):
+def slugify(s):
+
+    # Remove all non-word characters (everything except numbers and letters)
+    s = re.sub(r"[^\w\s]", '', s)
+
+    # Replace all runs of whitespace with a single dash
+    s = re.sub(r"\s+", '_', s)
+
+    return str(s)
+
+
+def check_nodes(*networks):
     
     isNone = [True for x in networks if x is None]
     if len(networks) - len(isNone) < 2:
@@ -332,9 +88,9 @@ def _check_nodes(*networks):
     return
 
 
-def _complete_networks(electrical_network,
-                       moorings_network,
-                       user_network):
+def complete_networks(electrical_network,
+                      moorings_network,
+                      user_network):
 
     # Determine which hierarchies/boms are available and create dummy
     # versions for any which are missing
@@ -484,16 +240,16 @@ def _complete_networks(electrical_network,
     return electrical_network, moorings_network, user_network
 
 
-def _combine_networks(electrical_network,
-                      moorings_network,
-                      user_network,
-                      electrical_data=None):
+def combine_networks(electrical_network,
+                     moorings_network,
+                     user_network,
+                     electrical_data=None):
     
     # Read in sub-system networks and consolidate into device- and array-level 
     # networks
-    _check_nodes(electrical_network,
-                 moorings_network,
-                 user_network)
+    check_nodes(electrical_network,
+                moorings_network,
+                user_network)
     
     # Create dictionary from electrical data
     if electrical_data is not None:
@@ -613,54 +369,7 @@ def _combine_networks(electrical_network,
     return array_hierarcy, device_hierachy
 
 
-def _get_indices(pool, label):
-        
-    labels, indices = find_all_labels(label, pool, partial_match=True)
-    if labels is None: return None
-    
-    subhub_indices = {}
-    subhub_names = [x[-1] for x in labels]
-    
-    for name, idx in zip(subhub_names, indices):
-        subhub_indices[name] = idx
-    
-    return subhub_indices
-
-
-def _get_curtailments(pool):
-    
-    hublist, _ = find_all_labels('device',
-                                 pool,
-                                 partial_match=True)
-    device_strings = find_strings(pool)
-    
-    curtailments = {}
-    device_names = [x[-1] for x in hublist]
-    
-    # Array
-    curtailments["array"] = device_names
-    
-    # Subhubs
-    if len(hublist[0]) == 3:
-        
-        subhub_names = set([x[1] for x in hublist])
-        
-        for name in subhub_names:
-            curtailments[name] = [x[-1] for x in hublist if x[1] == name]
-    
-    # devices
-    for name in device_names:
-        for string in device_strings:
-            
-            if name in string:
-                device_idx = string.index(name)
-                dev_curtails = string[device_idx:]
-                curtailments[name] = dev_curtails
-    
-    return curtailments
-
-
-def _build_pool(array_hierarcy, device_hierachy):
+def build_pool(array_hierarcy, device_hierachy):
     
     pool = {}
     array_link = Serial("array")
@@ -993,124 +702,6 @@ def _strip_dummy_k(compsids, kfactors):
     return idlist, kfactorlist
 
 
-def _set_component_failure_rates (pool,
-                                  dbdict,
-                                  severitylevel,
-                                  calcscenario,
-                                  use_kfactors=False):
-    
-    # For components with an id number look up respective failure rates 
-    # otherwise for designed components (i.e. shallow/gravity foundations, 
-    # direct embedment anchors and suction caissons) in addition to 
-    # grouted jointed use generic failure rate of 1.0x10^-4 failures 
-    # per annum (10 / 876 failures per 10^6 hours)
-    
-    # Note:
-    #  * If no data for a particular calculation scenario, failure rate 
-    #    defaults to mean value
-    #  * If no non-critical failure rate data is available use critical values
-    
-    def set_failure_rate(item, failure_rate, severitylevel, use_kfactors):
-        
-        if use_kfactors and item.kfactor is not None:
-            failure_rate *= item.kfactor
-        
-        item.set_failure_rate(failure_rate)
-        item.set_severity_level(severitylevel)
-        
-        return
-    
-    designed_comps = ["dummy",
-                      "n/a",
-                      "ideal",
-                      "gravity",
-                      "shallowfoundation",
-                      "suctioncaisson",
-                      "directembedment",
-                      "grout"]
-    
-    critical_key = 'failratecrit'
-    non_critical_key = 'failratenoncrit'
-    
-    if severitylevel == 'critical':
-        severity_key = critical_key
-        other_key = non_critical_key
-        other_severitylevel = 'noncritical'
-    elif severitylevel == 'noncritical':
-        severity_key = non_critical_key
-        other_key = critical_key
-        other_severitylevel = 'critical'
-    else:
-        err_str = ("Argument 'severitylevel' may only take values "
-                   "'critical' or 'noncritical'")
-        raise ValueError(err_str)
-    
-    lower_idx = 0
-    mean_idx = 1
-    upper_idx = 2
-    
-    if calcscenario == 'lower':
-        cs = lower_idx
-    elif calcscenario == 'mean':
-        cs = mean_idx
-    elif calcscenario == 'upper':
-        cs = upper_idx
-    else:
-        err_str = "Argument 'calcscenario' may only take values 0, 1, or 2"
-        raise ValueError(err_str)
-    
-    for item in pool.values():
-        
-        if not isinstance(item, Component):
-            continue
-        
-        if item.label in designed_comps:
-            item.set_failure_rate(10. / 876)
-            item.set_severity_level(severitylevel)
-            continue
-        
-        dbitem = deepcopy(dbdict[item.label]['item10'])
-        severity_failure_rates = dbitem[severity_key]
-        
-        if severity_failure_rates[cs] > 0.0:
-            failure_rate = severity_failure_rates[cs]
-            set_failure_rate(item, failure_rate, severitylevel, use_kfactors)
-            continue
-        
-        if severity_failure_rates[mean_idx] > 0.0:
-            failure_rate = severity_failure_rates[mean_idx]
-            set_failure_rate(item, failure_rate, severitylevel, use_kfactors)
-            continue
-        
-        other_failure_rates = dbitem[other_key]
-        
-        if other_failure_rates[cs] > 0.0:
-            
-            failure_rate = other_failure_rates[cs]
-            set_failure_rate(item,
-                             failure_rate,
-                             other_severitylevel,
-                             use_kfactors)
-            
-            continue
-        
-        if other_failure_rates[mean_idx] > 0.0:
-            
-            failure_rate = other_failure_rates[mean_idx]
-            set_failure_rate(item,
-                             failure_rate,
-                             other_severitylevel,
-                             use_kfactors)
-            
-            continue
-        
-        err_str = ("No failure rate data is set for component "
-                   "'{}'").format(item.label)
-        raise RuntimeError(err_str)
-    
-    return
-
-
 def _get_cable_kfactors(markers, data):
     
     m2km = lambda x: x / 1e3 
@@ -1124,7 +715,7 @@ def _get_cable_kfactors(markers, data):
         
         item_data = data[item]
         
-        if getattr(item_data, "Installation Type") in ["array", "export"]:
+        if getattr(item_data, "Installation_Type") in ["array", "export"]:
             kfactor = m2km(getattr(item_data, "Quantity"))
         else:
             kfactor = 1
